requirements.txt
requests
pandas
pymongo
python-dotenv
jupyter
beautifulsoup4
lxml
pycountry
matplotlib
jupyter-book
scipy
tqdm
